Renamed models -> utils. Changed corresponding imports from most files (still working on RhythmicEntry & PsetInstructor)
