+++ app/javascript/p_set/components/melodic_entry.js
@@ -3,10 +3,9 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import _ from 'lodash';
 import ReactAudioPlayer from 'react-audio-player';
-import { changeAudioPlayerState } from './helper';
 
 import VexflowComponent from './vexflow';
-import { formatKey, nextNonEmptyMeasure, prevNonEmptyMeasure, keyOptionToSignature, getVFScaleName } from '../lib/models';
+import { formatKey, nextNonEmptyMeasure, prevNonEmptyMeasure, keyOptionToSignature, getVFScaleName, changeAudioPlayerState } from '../lib/utils';
 
 const clefToOctave = (clef) => {
   if (clef === 'treble') {
